+++ CHANGELOG.md
@@ -7,6 +7,9 @@ timeline: true
 
 ## 1.1.x
 
+`2019-03-15`
+- :sunny: 新增 `ellipsis` 限制文字添加省略号 []()
+
 `2019-01-26`
 - :sunny: 新增 `randomNum` 产生指定范围的随机整数 []()
 - :sunny: 新增 `shuffle` 数组洗牌随机排序 []()